requirements.txt
fastapi
uvicorn
tensorflow-cpu
numpy
pretty_midi
librosa
pydantic
python-multipart

